% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_int.R
\name{.int_pseudo_sensitivity}
\alias{.int_pseudo_sensitivity}
\title{Generate pseudo-IPD and refit glmer for sensitivity analysis}
\usage{
.int_pseudo_sensitivity(data, input_spec, study_col, gamma, beta, rho, ...)
}
\description{
Uses gamma and beta from the aggregate model to generate pseudo-IPD,
then fits an individual-level glmer.
}
\keyword{internal}
