% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{log_prior_density}
\alias{log_prior_density}
\title{Evaluate log-prior density for a parameter vector}
\usage{
log_prior_density(theta, prior_obj)
}
\arguments{
\item{theta}{Named numeric vector of parameters.}

\item{prior_obj}{A \code{spima_prior} object.}
}
\value{
Log-density value (summed across independent priors).
}
\description{
Evaluate log-prior density for a parameter vector
}
