% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample from the joint prior}
\usage{
sample_prior(n, prior_obj)
}
\arguments{
\item{n}{Number of samples.}

\item{prior_obj}{A \code{spima_prior} object.}
}
\value{
A matrix with \code{n} rows and one column per prior.
}
\description{
Sample from the joint prior
}
