% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_bin.R
\name{spima_bin_observed_stats}
\alias{spima_bin_observed_stats}
\title{Compute Observed Summary Statistics for Binary Data}
\usage{
spima_bin_observed_stats(data, input_spec)
}
\arguments{
\item{data}{Original data frame per blueprint.}

\item{input_spec}{Column mapping.}
}
\value{
Named vector of observed log-ORs (or log-odds) with optional
  inverse-variance weights as attribute.
}
\description{
Compute Observed Summary Statistics for Binary Data
}
