% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cont.R
\name{spima_cont_observed_stats}
\alias{spima_cont_observed_stats}
\title{Compute Observed Summary Statistics for Continuous Data}
\usage{
spima_cont_observed_stats(data, input_spec)
}
\arguments{
\item{data}{Original data frame.}

\item{input_spec}{Column mapping.}
}
\value{
A list with \code{means} and \code{sds} (named vectors).
}
\description{
Compute Observed Summary Statistics for Continuous Data
}
