% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generic.R
\name{spima_generic_analyze}
\alias{spima_generic_analyze}
\title{Analyze Pseudo-Data for Generic Effect-Size}
\usage{
spima_generic_analyze(pseudo_ipd, input_spec)
}
\arguments{
\item{pseudo_ipd}{A data frame from \code{spima_bin_simulate}.}

\item{input_spec}{Column mapping.}
}
\value{
A list with \code{estimates = NULL},
  \code{summary_stats} (the effect-size vector), and
  \code{converged = TRUE}.
}
\description{
For the generic module, the "pseudo-IPD" is already the summary
statistics (a vector of effect sizes). This function simply passes
them through with converged = TRUE.
}
