% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_generic.R
\name{spima_generic_observed_stats}
\alias{spima_generic_observed_stats}
\title{Compute Observed Summary Statistics for Generic Effect-Size}
\usage{
spima_generic_observed_stats(data, input_spec)
}
\arguments{
\item{data}{Original data frame.}

\item{input_spec}{Column mapping.}
}
\value{
Named vector of effect sizes with \code{"weights"} attribute
  (inverse-variance: 1/sei^2).
}
\description{
Compute Observed Summary Statistics for Generic Effect-Size
}
